The animovement

The animovement is an extensible suite of R packages. It is a descendant of the fastverse.

Functions in the animovement Package

Functions to extend or reduce the number of packages in the animovement

animovement_extend()
animovement_detach()

Function to display conflicts for animovement packages (or any other attached packages)

animovement_conflicts()

Function to update animovement packages (and dependencies) and install (missing) packages

animovement_update()
animovement_install()

Utilities to retrieve the names of animovement packages (and dependencies), their update status and produce a situation report.

animovement_packages()
animovement_deps()
animovement_sitrep()

animovement Options

options(animovement.quiet = TRUE) will disable all automatic messages (including conflict reporting) when calling library(animovement), animovement_extend, animovement_update(install = TRUE) and animovement_install.

options(animovement.styling = FALSE) will disable all styling applied to text printed to the console.

options(animovement.extend = c(...)) can be set before calling library(animovement) to extend the animovement with some packages for the session. The same can be done with the animovement_extend function after library(animovement), which will also populate options("animovement.extend").

options(animovement.install = TRUE) can be set before library(animovement) to install any missing packages beforehand. See also animovement_install.

See also

Author

Maintainer: Mikkel Roald-Arbøl animovement.84w1m@passmail.com (ORCID)